WIP Added oauth capabilities
All of the code has been copied from:
https://github.com/lepture/example-oauth2-server

The oauth server is mostly done, the client isn't.

Things can't get started because I am not sure of how the views should
be created.

But as soon as we have a website, and some of the views get exposed,
things will become a lot clear.

Until then...