[hack] Login the user after creation
This is BAD as the API should never be used to login a user, and is only
temporary.

Once we have a better registeration process (one invoving email id
verification) this can be changed.

The other fix might be to remove ajax and use a form similar to the
login page, but that means the create_user code gets duplicated.