Update Users and Ideas models
The json property is now created by manually entering the keys, rather
than getting them from vars. The method is laborious but it gives us
total control over what we want to display.

User's password, email address are not displayed in the json now, just
as they shouldn't be.

Also added a user property to ideas which displays basic data.