Added a RAW query to fetch ideas
The code works, and returns a row of tuples. BUT...

We've been using objects of all types in our code, which is almost
impossible to be replaced by queries now.

I tried creating a function that could convert the data to json, but
then we would lose some or the other attribute.

I can't think of a way that doesn't involve some sort of hackery.

:/