Only show_user doesnt work.
uuids are really creating problems.

When a new idea is created that time the url contains the hyphens.

At that time everythin works with that idea only.

But once the server is closed and restarted that time all the hyphens are
absent in the url and show_idea stops working.

Additionally, avatar pic is not loading, profile form is not being submitted

Also, get_idea() in the API still doesn't work.